fix(editor): keep whitespace between adjacent blocks, hoist the markup regex
Self-review follow-ups on the line-accounting change:
- the adjacent-tag pass dropped the horizontal whitespace separating two
  blocks while the file's header promises it only ADDS. It is carried
  through now, and the "never removes characters" test covers the spaced
  and tabbed forms (mutation-checked: reverting the replacement goes red).
- the structural-markup regex is a constant; build it once, not per call.
- expectedTotalLines is an integer on the route, matching the SDK schema —
  a fractional value could never equal a line count, so it would have
  refused every edit that supplied one.

+++ apps/web/src/app/api/mcp/documents/route.ts
@@ -145,7 +145,7 @@ const lineOperationSchema = z.object({
   // Optional staleness guard for write operations: the totalLines the caller
   // read before addressing this edit. Supplying it turns "the document grew
   // since I looked" from a silent partial overwrite into a 409.
-  expectedTotalLines: z.number().min(0).optional(),
+  expectedTotalLines: z.number().int().min(0).optional(),
 });
 
 export async function POST(req: NextRequest) {
